requirements.txt
mediapipe
opencv-python
sounddevice
scipy
openai-whisper
torch
